initial backend refactor
