+++ backend/RIPmain/readme.md
@@ -1,0 +1,1 @@
+# This is the old code used in the hackathon before the architectural refactor
