+++ backend/RIPmain/requirements.in
@@ -1,0 +1,78 @@
+absl-py
+astunparse
+bidict
+blinker
+certifi
+charset-normalizer
+click
+cvlib
+distlib
+dlib
+face-recognition
+face_recognition_models
+filelock
+Flask
+Flask-Cors
+Flask-SocketIO
+flatbuffers
+fsspec
+gast
+google-pasta
+grpcio
+h11
+h5py
+idna
+imageio
+imutils
+iniconfig
+itsdangerous
+Jinja2
+keras
+libclang
+Markdown
+markdown-it-py
+MarkupSafe
+mdurl
+ml-dtypes
+mpmath
+namex
+networkx
+numpy
+opencv-python
+opencv-python-headless
+opt_einsum
+optree
+packaging
+pillow
+platformdirs
+pluggy
+progressbar
+protobuf
+Pygments
+pytest
+python-engineio
+python-socketio
+requests
+rich
+scipy
+setuptools
+simple-websocket
+six
+sympy
+tensorboard
+tensorboard-data-server
+tensorflow
+termcolor
+torch
+torchvision
+typing_extensions
+urllib3
+ultralytics
+uWSGI
+virtualenv
+Werkzeug
+wheel
+wrapt
+wsproto
+matplotlib
+ultralytics
