+++ backend/requirements.in
@@ -75,4 +75,4 @@ wheel
 wrapt
 wsproto
 matplotlib
-time
+ultralytics
